+++ r-package/grf/bindings/RegressionForestBindings.cpp
@@ -13,18 +13,18 @@
 Rcpp::List regression_train(Rcpp::NumericMatrix input_data,
                             Eigen::SparseMatrix<double> sparse_input_data,
                             size_t outcome_index,
-                            unsigned int mtry,
-                            unsigned int num_trees,
-                            unsigned int num_threads,
-                            unsigned int min_node_size,
+                            uint mtry,
+                            uint num_trees,
+                            uint num_threads,
+                            uint min_node_size,
                             double sample_fraction,
-                            unsigned int seed,
+                            uint seed,
                             bool honesty,
-                            unsigned int ci_group_size,
+                            uint ci_group_size,
                             double alpha,
                             double imbalance_penalty,
-                            std::vector<uint> clusters,
-                            unsigned int samples_per_cluster) {
+                            std::vector<size_t> clusters,
+                            uint samples_per_cluster) {
   ForestTrainer trainer = ForestTrainers::regression_trainer(outcome_index - 1);
 
   Data* data = RcppUtilities::convert_data(input_data, sparse_input_data);
@@ -44,8 +44,8 @@ Rcpp::List regression_train(Rcpp::NumericMatrix input_data,
 Rcpp::List regression_predict(Rcpp::List forest_object,
                               Rcpp::NumericMatrix input_data,
                               Eigen::SparseMatrix<double> sparse_input_data,
-                              unsigned int num_threads,
-                              unsigned int ci_group_size) {
+                              uint num_threads,
+                              uint ci_group_size) {
   Data* data = RcppUtilities::convert_data(input_data, sparse_input_data);
   Forest forest = RcppUtilities::deserialize_forest(
       forest_object[RcppUtilities::SERIALIZED_FOREST_KEY]);
@@ -62,8 +62,8 @@ Rcpp::List regression_predict(Rcpp::List forest_object,
 Rcpp::List regression_predict_oob(Rcpp::List forest_object,
                                   Rcpp::NumericMatrix input_data,
                                   Eigen::SparseMatrix<double> sparse_input_data,
-                                  unsigned int num_threads,
-                                  unsigned int ci_group_size) {
+                                  uint num_threads,
+                                  uint ci_group_size) {
   Data* data = RcppUtilities::convert_data(input_data, sparse_input_data);
   Forest forest = RcppUtilities::deserialize_forest(
       forest_object[RcppUtilities::SERIALIZED_FOREST_KEY]);
